fix: minor renames
