+++ src/opi/output/core.py
@@ -16,7 +16,7 @@ from pydantic import StrictInt, StrictStr
 from opi.execution.core import Runner
 from opi.input.structures import Atom, Coordinates, Structure
 from opi.output.cube import CubeOutput
-from opi.output.fcidump_parser import Fcidump
+from opi.output.fcidump import Fcidump
 from opi.output.gbw_suffix import GbwSuffix
 from opi.output.grepper.recipes import (
     get_error_message,
@@ -2746,8 +2746,8 @@ class Output:
 
     def get_fcidump(self) -> Fcidump | None:
         """
-        Parse the fcidump file generated by ORCA and return its data in the `Fcidump` data class.
-        The fcidump file has to be generated by the ORCA job and cannot be generated on-the-fly after the calculation.
+        Parse the FCIDUMP file generated by ORCA and return its data in the `Fcidump` data class.
+        The FCIDUMP file has to be generated by the ORCA job and cannot be generated on-the-fly after the calculation.
         To generate FCIDUMP files, set the following options:
         ```
         %output
@@ -2759,14 +2759,14 @@ class Output:
         Returns
         -------
         fcidump_data: Fcidump | None
-            The parsed fcidump data or None if the file is not present or could not be parsed.
+            The parsed FCIDUMP data or None if the file is not present or could not be parsed.
         """
 
-        # > Get path to the fcidump file
+        # > Get path to the FCIDUMP file
         fci_path = self.get_outfile().with_suffix(".fcidump")
 
         # > If there is no file we return None
         if not fci_path.is_file():
             return None
 
-        return Fcidump.parse_fcidump(fci_path)
+        return Fcidump.from_file(fci_path)
